+++ src/Endpoint2/Configuration/AuditErrorConfiguration.cs
