+++ src/Endpoint2/Configuration/PersistenceConfiguration.cs
@@ -6,12 +6,9 @@ public class PersistenceConfiguration : INeedInitialization
 {
     public void Customize(EndpointConfiguration configuration)
     {
-        configuration.EnableInstallers();
         var persistence = configuration.UsePersistence<SqlPersistence>();
         persistence.SubscriptionSettings().DisableCache();
-        persistence.TablePrefix("Endpoint_");
-        var dialect = persistence.SqlDialect<SqlDialect.MsSqlServer>();
-        //dialect.Schema("OrderContext");
-        persistence.ConnectionBuilder(connectionBuilder: () => new SqlConnection("Server=.;Database=nservicebus;Trusted_Connection=True;"));
+        persistence.SqlDialect<SqlDialect.MsSqlServer>();
+        persistence.ConnectionBuilder(() => new SqlConnection("Server=.;Database=nservicebus;Trusted_Connection=True;"));
     }
 }
