+++ src/Endpoint2/Configuration/SerializationConfiguration.cs
