+++ src/Endpoint2/Configuration/TransportConfiguration.cs
@@ -4,6 +4,6 @@ class TransportConfiguration : INeedInitialization
 {
     public void Customize(EndpointConfiguration endpointConfiguration)
     {
-        var transport = endpointConfiguration.UseTransport<MsmqTransport>();
+        endpointConfiguration.UseTransport<MsmqTransport>();
     }
 }
