--- src/Infra.NServiceBus/Recovery/RecoveryConfiguration.cs
@@ -1,22 +1,0 @@
-﻿using System;
-using NServiceBus;
-
-class RecoveryConfiguration : INeedInitialization
-{
-    public void Customize(EndpointConfiguration configuration)
-    {
-        var recoverability = configuration.Recoverability();
-        recoverability.DisableLegacyRetriesSatellite();
-        recoverability.Immediate(
-            customizations: immediate =>
-            {
-                immediate.NumberOfRetries(0);
-            });
-        recoverability.Delayed(
-            customizations: delayed =>
-            {
-                var numberOfRetries = delayed.NumberOfRetries(0);
-                numberOfRetries.TimeIncrease(TimeSpan.FromSeconds(30));
-            });
-    }
-}
