--- src/Infra.NServiceBus/SanitizationStrategy.cs
@@ -1,16 +1,0 @@
-﻿using NServiceBus;
-using NServiceBus.Transport.AzureServiceBus;
-
-class SanitizationStrategy : ISanitizationStrategy
-{
-    public string Sanitize(string entityPathOrName, EntityType entityType)
-    {
-        if (entityType == EntityType.Subscription)
-        {
-            var len = 50 >= entityPathOrName.Length ? 0 : entityPathOrName.Length - 50;
-            return entityPathOrName.Substring(len);
-        }
-
-        return entityPathOrName;
-    }
-}
