--- src/Infra.NServiceBus/packages.config
@@ -1,22 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="EntityFramework" version="6.2.0" targetFramework="net452" />
-  <package id="Iesi.Collections" version="4.0.3" targetFramework="net452" />
-  <package id="Microsoft.Azure.KeyVault.Core" version="1.0.0" targetFramework="net452" />
-  <package id="Microsoft.Data.Edm" version="5.8.3" targetFramework="net452" />
-  <package id="Microsoft.Data.OData" version="5.8.3" targetFramework="net452" />
-  <package id="Microsoft.Data.Services.Client" version="5.8.3" targetFramework="net452" />
-  <package id="Microsoft.WindowsAzure.ConfigurationManager" version="3.2.3" targetFramework="net452" />
-  <package id="Newtonsoft.Json" version="11.0.2" targetFramework="net46" />
-  <package id="NServiceBus" version="6.4.3" targetFramework="net452" />
-  <package id="NServiceBus.Azure.Transports.WindowsAzureServiceBus" version="7.2.11" targetFramework="net452" />
-  <package id="NServiceBus.Hosting.Azure" version="7.1.3" targetFramework="net452" />
-  <package id="NServiceBus.Hosting.Azure.HostProcess" version="7.1.3" targetFramework="net452" />
-  <package id="NServiceBus.Metrics" version="2.0.0" targetFramework="net452" />
-  <package id="NServiceBus.Metrics.ServiceControl" version="2.0.0" targetFramework="net452" />
-  <package id="NServiceBus.Persistence.Sql" version="3.1.0" targetFramework="net46" />
-  <package id="NServiceBus.Persistence.Sql.MsBuild" version="3.1.0" targetFramework="net46" developmentDependency="true" />
-  <package id="System.Spatial" version="5.8.3" targetFramework="net452" />
-  <package id="WindowsAzure.ServiceBus" version="4.1.6" targetFramework="net452" />
-  <package id="WindowsAzure.Storage" version="8.7.0" targetFramework="net452" />
-</packages>
