+++ src/NsbAsbOutboxSqlPersistenceEf-SupportSample.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 15.0.27428.2027
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain", "Domain\Domain.csproj", "{11B8C825-077F-4726-A55C-842277850A32}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Infra.NServiceBus", "Infra.NServiceBus\Infra.NServiceBus.csproj", "{1C31F37C-053B-478A-A0FA-AECC0408D4FD}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Endpoint2", "Endpoint2\Endpoint2.csproj", "{0F372666-A717-4CB9-B459-4509CEC47C55}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "DatabaseScripts", "DatabaseScripts", "{9799F327-BFD6-4F30-9A25-4880A8203767}"
@@ -35,16 +33,6 @@ Global
 		{11B8C825-077F-4726-A55C-842277850A32}.Release|Any CPU.Build.0 = Release|Any CPU
 		{11B8C825-077F-4726-A55C-842277850A32}.test4-ea|Any CPU.ActiveCfg = test4-ea|Any CPU
 		{11B8C825-077F-4726-A55C-842277850A32}.test4-ea|Any CPU.Build.0 = test4-ea|Any CPU
-		{1C31F37C-053B-478A-A0FA-AECC0408D4FD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1C31F37C-053B-478A-A0FA-AECC0408D4FD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1C31F37C-053B-478A-A0FA-AECC0408D4FD}.dit-ea|Any CPU.ActiveCfg = dit-ea|Any CPU
-		{1C31F37C-053B-478A-A0FA-AECC0408D4FD}.dit-ea|Any CPU.Build.0 = dit-ea|Any CPU
-		{1C31F37C-053B-478A-A0FA-AECC0408D4FD}.Mat|Any CPU.ActiveCfg = Mat|Any CPU
-		{1C31F37C-053B-478A-A0FA-AECC0408D4FD}.Mat|Any CPU.Build.0 = Mat|Any CPU
-		{1C31F37C-053B-478A-A0FA-AECC0408D4FD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1C31F37C-053B-478A-A0FA-AECC0408D4FD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1C31F37C-053B-478A-A0FA-AECC0408D4FD}.test4-ea|Any CPU.ActiveCfg = test4-ea|Any CPU
-		{1C31F37C-053B-478A-A0FA-AECC0408D4FD}.test4-ea|Any CPU.Build.0 = test4-ea|Any CPU
 		{0F372666-A717-4CB9-B459-4509CEC47C55}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{0F372666-A717-4CB9-B459-4509CEC47C55}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0F372666-A717-4CB9-B459-4509CEC47C55}.dit-ea|Any CPU.ActiveCfg = Debug|Any CPU
